Re-disable touch selection.
This is nice to have disabled, otherwise when you're hodling your finger
down waiting to start, my browser (Chrome on Android) starts to select
the text on the screen, which is definitely not what I want.

This code was commented out in
http://github.com/cubing/timer/commit/9e7a49816d6865d4b8a9a2e7d3e4a582318252d7,
but it's not clear why. I lightly tested this change, and listing times
does still seem to work.

After some discussion with @lgarron, we decided to refine the css
selector a bit so we only disable text selection on the `.timer`
element. This is nice because we do want to let users select other
things on the page (such as the scramble).